DAT-16380 (#20)
* Fix docs.

* Implement specific workflowId for downloading artifacts

Added a parameter for workflowId in the downloadArtifact method. Previously, the method was using the global workflowId for both the main workflow and the Pro workflow, which would cause issues if these workflows had different IDs. Now, the Pro workflow uses its own specific workflowId to download artifacts, ensuring the correct artifact is downloaded for each respective workflow.