Add support for different workflow IDs in InstallSnapshotMojo
Add a new parameter 'proWorkflowId' to distinguish between workflows in liquibase/liquibase-pro and other repositories. Previously, all repositories were assumed to use the same workflow ID, but this is not always the case. This update allows the correct workflow ID to be used per repository, enhancing the flexibility and accuracy of the program.